Migrate form PHPStan to Mago for code analysis

+++ src/Resolver/GeoLite2LocationResolver.php
@@ -47,8 +47,8 @@ class GeoLite2LocationResolver implements IpLocationResolverInterface
             $city->country->name ?? '',
             $city->subdivisions[0]->name ?? '',
             $city->city->name ?? '',
-            (float) ($city->location->latitude ?? 0.0),
-            (float) ($city->location->longitude ?? 0.0),
+            $city->location->latitude ?? 0.0,
+            $city->location->longitude ?? 0.0,
             $city->location->timeZone ?? '',
         );
     }
